preferences::PreferencesMap

type PreferencesMap<T = String> = HashMap<String, T>;

Generic key-value store for user data.

This is actually a wrapper type around std::collections::HashMap<String, T> (with T defaulting to String), so use the HashMap API methods to access and change user data in memory.

To save or load user data, use the methods defined for PreferencesTrait, which will be automatically implemented for PreferencesMap<T> as long as T is serializable. (See the module documentation for examples and more details.)